Bring voer getFileText and rename methods
